pyproject.toml
[tool.poetry]
name = "sept-2023-drf"
version = "0.1.0"
description = ""
authors = ["akiselev1983 <akiselev.brd@gmail.con>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
mysqlclient = "^2.2.4"



[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
